+++ lib/mongoid/relations/bindings/referenced/in.rb
@@ -27,9 +27,7 @@ module Mongoid # :nodoc:
                   base.you_must(metadata.inverse_type_setter, target.class.model_name)
                 end
                 if inverse
-                  inverse_metadata = metadata.inverse_metadata(target)
-                  if inverse_metadata != metadata && !inverse_metadata.nil?
-                    base.metadata = inverse_metadata
+                  if set_base_metadata
                     if base.referenced_many?
                       target.send(inverse).push(base) unless Mongoid.identity_map_enabled?
                     else
@@ -59,6 +57,7 @@ module Mongoid # :nodoc:
                   base.you_must(metadata.inverse_type_setter, nil)
                 end
                 if inverse
+                  set_base_metadata
                   if base.referenced_many?
                     target.send(inverse).delete(base)
                   else
@@ -69,6 +68,27 @@ module Mongoid # :nodoc:
             end
           end
           alias :unbind_one :unbind
+
+          private
+
+          # Ensure that the metadata on the base is correct, for the cases
+          # where we have multiple belongs to definitions and were are setting
+          # different parents in memory in order.
+          #
+          # @api private
+          #
+          # @example Set the base metadata.
+          #   binding.set_base_metadata
+          #
+          # @return [ true, false ] If the metadata changed.
+          #
+          # @since 2.4.4
+          def set_base_metadata
+            inverse_metadata = metadata.inverse_metadata(target)
+            if inverse_metadata != metadata && !inverse_metadata.nil?
+              base.metadata = inverse_metadata
+            end
+          end
         end
       end
     end